fate/ffmpeg: remove test for fix_sub_duration_heartbeat
After the full ffmpeg CLI multithreading changes went in, this
test started depending on how far the input side read and decoded
the input compared to how quickly the output encoded things, causing
spurious failures on the CI. If the input gets further, subtitles'
end points are more likely to be found out and thus their final form
would be passed on to muxer as-is. On the other hand, if the input
does not get that far, an intermediate split subtitle is output
instead, with the duration ending at the point of a random access
point packet that caused the state of the ongoing subtitle to be
checked.

To my knowledge all of the failures have so far been valid correct
results, but unfortunately FATE's built in checks mostly consist of
whether there is a difference against an exact result. In the case
of this test, what would need to be checked is that the same subtitle
is indeed shown from time A to time B, but it may be split into
multiple parts (limited by size of GOPs output during the test,
which is defined in it). Alternatively, this test could be fixed
by making it possible to define how far the input is allowed to
read and decode compared to the output. That is currently not
something implemented in the project, either.

All in all, I dislike lowering the test coverage of code, but this
removal will at least make people less aggravated about a test
spuriously failing.