lavu/crc: add a function pointer based API (PR #21108)
PR #21108 opened by Lynne
URL: https://code.ffmpeg.org/FFmpeg/FFmpeg/pulls/21108
Patch URL: https://code.ffmpeg.org/FFmpeg/FFmpeg/pulls/21108.patch

This commit adds a new function pointer based API for CRC.
The CRC code is some of the oldest in the entire codebase. Despite
CRC instructions existing on most CPUs back when it was written, it
did not implement support for them, nor even 64-bit CRCs.

Our TTA decoder even NIHs its own 64-bit CRC just because the
API is simply not capable of it!

Attempts to use architecture-specific CRCs remain blocked on this.

The new API is much simpler, implementation-wise.
It was copied verbatim from the lavu/tx API.

There will be no support for hardcoded tables: users are already expected
to keep state (the function pointer), and they'll have to init a context.

We'll make the old code a wrapper for the new if this is accepted,
and we'll get rid of hardcoded tables to save space.

Ref: #20751

>From 5dfa403995e97f1aba251b19d1d39cca6bf0a6f4 Mon Sep 17 00:00:00 2001
From: Lynne <dev@lynne.ee>
Date: Fri, 5 Dec 2025 10:07:59 +0100
Subject: [PATCH] lavu/crc: add a function pointer based API

This commit adds a new function pointer based API for CRC.
The CRC code is some of the oldest in the entire codebase. Despite
CRC instructions existing on most CPUs back when it was written, it
did not implement support for them, nor even 64-bit CRCs.

Our TTA decoder even NIHs its own 64-bit CRC just because the
API is simply not capable of it!

Attempts to use architecture-specific CRCs remain blocked on this.

The new API is much simpler, implementation-wise.
It was copied verbatim from the lavu/tx API.

There will be no support for hardcoded tables: users are already expected
to keep state (the function pointer), and they'll have to init a context.

We'll make the old code a wrapper for the new if this is accepted,
and we'll get rid of hardcoded tables to save space.