configure: set PT_GNU_STACK to 1M
This requests stack size. With Musl libc, the default thread stack size
is 128KB, which is not enough for all the code.

One example is VVC's put_scaled() which uses ~320KB on stack. While this
would be good to improve this and other instances where stack usage is
excessive in the meantime let's request size as needed.

Signed-off-by: Kacper Michajłow <kasper93@gmail.com>