avformat/mov: fixed bug in mov_merge_tts_data (PR #21234)
PR #21234 opened by anders-mjoll
URL: https://code.ffmpeg.org/FFmpeg/FFmpeg/pulls/21234
Patch URL: https://code.ffmpeg.org/FFmpeg/FFmpeg/pulls/21234.patch

After upgrading to FFmpeg 8.0 (from FFmpeg 5.1.2) we got some customers reporting that FFmpeg consumed all the memory of the machine and was eventually killed by the OS. I was able to track it down to the MOV demuxer not producing any audio packets and the filter graph just caching up all the video frames from the 70 GB MOV file. This in turn appears to be because the index is never built for the audio stream. The reason it doesn't get built is because it fails the following check in `mov_merge_tts_data`:

```
  if (!sc->sample_count || sc->sample_count >= UINT_MAX / sizeof(*sc->tts_data))
        return -1;
```

This happens because this MOV file contains PCM and in that case it appears that the `sc->sample_count` is equal to the actual number of PCM samples not the number of entries in the `stts` and `ctts` tables. This PR fixes this issue by first obtaining the actual number of entries required by the ctts and stts tables.

Unfortunately the file that triggers this problem is 70 GB and also contains sensitive customer data so I cannot provide a sample to reproduce the problem.

>From 0c2d59273ba09c73a97fee64e86aa190367228b8 Mon Sep 17 00:00:00 2001
From: Anders Rein <anders@onemimir.com>
Date: Thu, 18 Dec 2025 18:59:17 +0100
Subject: [PATCH] avformat/mov: fixed bug in mov_merge_tts_data

Calculate the actual required size for the tts_data array instead of
assuming sc->sample_count reflect the number of entries in the stts/ctts
tables. This assumption does not hold for certain MOV files with PCM
audio.