avformat/wavenc: use RF64 when needed
Currently if you encode a large wav, ffmpeg will hint after the fact
that the resultant file is corrupted, because traditional wav cannot
handle lengths greater than 32-bits. This isn't very useful; nobody
benefits from getting garbage files.

You can manually work around this by adding `-rf64 always`, which most
players have support for. Most people don't remember to do this until
after the fact when their file is corrupted, or they don't figure it out
at all and wind up using the w64 container instead.

And so that's what `-rf64 auto` is for. It uses the larger format when
needed, and if not, uses the traditional wav that is probably more
compatible. The result of using `-rf64 auto` is that you can add it to
every command line -- and should add it to every command line -- to get
either a normal small file, or a non-corrupt large file.

This is a very sensible default to have on, rather than just producing
corrupt files and having users scrambling for solutions, and then having
to do a potentially expensive reencode after. With `-rf64 auto` on by
default, the user always gets a readable good file. And for users who
sometimes want corrupt files, there still exists `-rf64 never` that can
be enabled.